+++ src/arctix/dataset/multithumos.py
@@ -649,27 +649,27 @@ def to_array(frame: pl.DataFrame) -> dict[str, np.ndarray]:
                 dtype=int,
                 max_len=mask.shape[1],
                 padded_value=-1,
-            ).astype(int),
+            ),
             mask=mask,
         ),
         Column.END_TIME: np.ma.masked_array(
             data=convert_sequences_to_array(
                 groups.get_column(Column.END_TIME).to_list(),
-                dtype=float,
+                dtype=np.float64,
                 max_len=mask.shape[1],
                 padded_value=-1.0,
-            ).astype(float),
+            ),
             mask=mask,
         ),
-        Column.SEQUENCE_LENGTH: lengths.astype(int),
+        Column.SEQUENCE_LENGTH: lengths.astype(np.int64),
         Column.SPLIT: groups.get_column(Column.SPLIT).to_numpy().astype(str),
         Column.START_TIME: np.ma.masked_array(
             data=convert_sequences_to_array(
                 groups.get_column(Column.START_TIME).to_list(),
-                dtype=float,
+                dtype=np.float64,
                 max_len=mask.shape[1],
                 padded_value=-1.0,
-            ).astype(float),
+            ),
             mask=mask,
         ),
     }
